split upload to seperate
