+++ app.py
@@ -1,30 +1,13 @@
-import os
 from flask import Flask, request, render_template, jsonify, make_response
-from werkzeug.utils import secure_filename
+import environment
 
 app = Flask(__name__)
-
-UPLOAD_FOLDER = 'storage'
-ALLOWED_EXTENSIONS = {'jpg', 'png', 'jpeg'}
-
-def allowed_file(filename):
-    return '.' in filename and filename.rsplit('.', 1)[1].lower() in ALLOWED_EXTENSIONS
 
 @app.route("/", methods=['GET', 'POST'])
 def home():
     if request.method == 'POST':
         uploaded_file = request.files['file-to-upload']
-
-        # If the user does not select a file, the browser submits an
-        # empty file without a filename.
-        if uploaded_file.filename == '':
-            return 'No selected file'
-
-        if not allowed_file(uploaded_file.filename):
-            return 'FILE NOT ALLOWED!'
-
-        filename = secure_filename(uploaded_file.filename)
-        uploaded_file.save(os.path.join(UPLOAD_FOLDER, filename))
+        environment.upload(uploaded_file)
 
     return render_template('index.html')
